```
benchmark           old ns/op     new ns/op     delta
BenchmarkYOLO-4     733923927     734250451     +0.04%

benchmark           old allocs     new allocs     delta
BenchmarkYOLO-4     19045          11570          -39.25%

benchmark           old bytes     new bytes     delta
BenchmarkYOLO-4     184515390     184207324     -0.17%
```